+++ CHANGELOG.md
@@ -89,9 +89,10 @@
 - Update copyright and license notices. (#254)
 - Migrate puml2links.sh to zx. (#260)
 - Make Poco Boost ITs runnable with "native" mode. (#271)
-- Add Slither config. Run single contract Slither analysis on CI. (#244)
+- Add Slither config. Run single contract Slither analysis on CI. (#244, #275)
 - Remove useless files. (#273)
 - Merge settings and v8Settings using object spread. (#274)
+- Create a hardhat network close to the iExec Bellecour blockchain. (#276)
 
 ## v5.4.2
 - Use latest Nethermind base image `nethermindeth/nethermind:iexec`